feat: add tinymce + refactor part 1
