+++ package.json
@@ -26,6 +26,7 @@
     "@radix-ui/react-tabs": "^1.0.4",
     "@radix-ui/react-toast": "^1.1.4",
     "@radix-ui/react-tooltip": "^1.0.6",
+    "@tinymce/tinymce-react": "^4.3.0",
     "@web3auth/modal": "^6.1.7",
     "@web3auth/openlogin-adapter": "^6.1.7",
     "@web3auth/web3auth-wagmi-connector": "^4.0.1",
