+++ pages/project/create/index.tsx
@@ -1,7 +1,8 @@
 import { AuthContext } from "@/components/AuthProvider";
 import { BenefitsFields } from "@/components/BenefitsFields";
+import { ProjectSection } from "@/components/CreateProjectPage/ProjectSection";
 import { MilestoneFields } from "@/components/MilestoneFields";
-import { TeamFields } from "@/components/TeamFields";
+import { TeamFields } from "@/components/CreateProjectPage/TeamFields";
 import { chains } from "@/components/Web3Provider";
 import { Button } from "@/components/ui/button";
 import { Calendar } from "@/components/ui/calendar";
@@ -30,13 +31,6 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from "@/components/ui/popover";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 import { Textarea } from "@/components/ui/textarea";
 import {
   Tooltip,
@@ -48,7 +42,6 @@ import { useToast } from "@/components/ui/use-toast";
 import { VIMEO_REGEX, YOUTUBE_REGEX } from "@/constants/regex";
 import { generateVideoThumbnail } from "@/lib/generateVideoThumbnail";
 import { cn } from "@/lib/utils";
-import { Brief } from "@/types/mongo";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { format } from "date-fns";
 import { CalendarIcon } from "lucide-react";
@@ -59,16 +52,18 @@ import { useForm } from "react-hook-form";
 import { isAddress } from "viem";
 import { useAccount, useEnsAddress, useMutation, useQuery } from "wagmi";
 import * as z from "zod";
+import { TeamSection } from "@/components/CreateProjectPage/TeamSection";
+import { SupportSection } from "@/components/CreateProjectPage/SupportSection";
 
 interface CreateProjectValues
-  extends Omit<z.infer<typeof formSchema>, "mint_end_date" | "tags"> {
+  extends Omit<z.infer<typeof createFormSchema>, "mint_end_date" | "tags"> {
   mint_end_date: string;
   tags: string[];
 }
 
 async function createProject(
   idToken: string,
-  values: z.infer<typeof formSchema>
+  values: z.infer<typeof createFormSchema>
 ) {
   const finalValues: CreateProjectValues = {
     ...values,
@@ -118,7 +113,7 @@ async function getCheckoutLink(
   return "";
 }
 
-const formSchema = z.object({
+export const createFormSchema = z.object({
   title: z.string().min(1, { message: "Title is required" }),
   description: z.string().min(1, { message: "Description is required" }),
   video_url: z
@@ -183,8 +178,8 @@ export default function ProjectForm() {
   const { address } = useAccount();
   const { idToken } = useContext(AuthContext);
 
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
+  const form = useForm<z.infer<typeof createFormSchema>>({
+    resolver: zodResolver(createFormSchema),
     mode: "onBlur",
     defaultValues: {
       title: "",
@@ -266,7 +261,7 @@ export default function ProjectForm() {
 
   console.log({ checkoutLink });
 
-  async function onSubmit(values: z.infer<typeof formSchema>) {
+  async function onSubmit(values: z.infer<typeof createFormSchema>) {
     // print form errors
     if (errors) {
       console.error({ errors, values });
@@ -315,271 +310,9 @@ export default function ProjectForm() {
         // @ts-expect-error For netlify forms
         netlify="true"
       >
-        <div className="border border-slate-200 rounded p-10 mb-10">
-          <h1 className="font-base">The Project</h1>
-          <p className="form-subheading">
-            {"Hey there future maker, what's your project?"}
-          </p>
-          <hr className="border-b-1 border-slate-200 my-8" />
-          <div className="grid grid-cols-2 gap-10">
-            <div className="col-span-1 pr-4">
-              <h2 className="text-xl font-base">Basic Info</h2>
-              <p>
-                Write a Clear and Concise Title and Subtitle for Your Project
-                <br />
-                <br />
-                Your project title and subtitle will appear on your project and
-                pre-launch pages, as well as in category pages, search results,
-                and emails we send to our community. Make sure they accurately
-                represent your project and are easy to understand for potential
-                supporters.
-              </p>
-            </div>
-            <div className="col-span-1">
-              <FormField
-                control={control}
-                name="title"
-                render={({ field }) => (
-                  <FormItem className="pb-4">
-                    <FormLabel>What is the name of your project?</FormLabel>
-                    <FormControl>
-                      <Input {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={control}
-                name="description"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Describe your idea in a sentence</FormLabel>
-                    <FormDescription className="text-xs">
-                      This will be featured on homepage alongside your video
-                    </FormDescription>
-                    <FormControl>
-                      <Textarea {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </div>
-          </div>
-          <hr className="border-b-1 border-slate-200 my-8" />
-          <div className="grid grid-cols-2 gap-10">
-            <div className="col-span-1 pr-4">
-              <h2 className="text-xl font-base">Summary</h2>
-              <p>
-                Please provide a brief summary that will motivate supporters to
-                believe in your vision. Be genuine rather than polished!
-                <br />
-                <br />
-                Explain what you aim to achieve with the funding, how you intend
-                to accomplish it, who you are, and why this project is important
-                to you. Demonstrations and step-by-step guides are highly
-                effective!
-              </p>
-            </div>
-            <div className="col-span-1">
-              <FormField
-                control={control}
-                name="video_url"
-                render={({ field }) => (
-                  <FormItem className="pb-4">
-                    <FormControl>
-                      <Input {...field} />
-                    </FormControl>
-                    <FormDescription>
-                      Share a brief 3-minute video through a URL (e.g. Vimeo or
-                      YouTube) introducing your vision for a better future.
-                    </FormDescription>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={control}
-                name="tldr"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Project TLDR</FormLabel>
-                    <FormControl>
-                      <Textarea {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </div>
-          </div>
-          <hr className="border-b-1 border-slate-200 my-8" />
-          <div className="grid grid-cols-2 gap-10">
-            <div className="col-span-1 pr-4">
-              <h2 className="text-xl font-base">Video Image</h2>
-              <p>
-                This image is taken from the thumbnail of your uploaded video.
-                <br />
-                <br />
-                This will appear for collectors in their wallet and on their
-                profile.
-              </p>
-            </div>
-            <div className="col-span-1 space-y-2">
-              <FormField
-                control={control}
-                name="video_image"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Enter a YouTube or Vimeo URL</FormLabel>
-                    <FormControl>
-                      <Input {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              {video_image !== "" && (
-                <img src={generateVideoThumbnail(video_image)} />
-              )}
-            </div>
-          </div>
-          <hr className="border-b-1 border-slate-200 my-8" />
-          <div className="grid grid-cols-2 gap-10">
-            <div className="col-span-1 pr-4">
-              <h2 className="text-xl font-base">Inspiration</h2>
-              <p>
-                {
-                  "Choose a brief that inspires a playful future, or select one of our partner briefs and explain why you're building it. We'll use this to communicate your vision in any email newsletters, interviews or social campaigns."
-                }
-              </p>
-            </div>
-            <div className="col-span-1">
-              <FormField
-                control={control}
-                name="brief"
-                render={({ field }) => (
-                  <FormItem className="pb-4">
-                    <FormLabel>Select a brief you are answering:</FormLabel>
-                    <Select
-                      onValueChange={field.onChange}
-                      defaultValue={field.value}
-                    >
-                      <FormControl>
-                        <SelectTrigger>
-                          <SelectValue placeholder="Select a Brief" />
-                        </SelectTrigger>
-                      </FormControl>
-                      <SelectContent>
-                        {Object.values(Brief).map((brief) => (
-                          <SelectItem key={brief} value={brief}>
-                            {brief}
-                          </SelectItem>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={control}
-                name="inspiration"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>
-                      What was the inspiration for this idea?
-                    </FormLabel>
-                    <FormControl>
-                      <Textarea {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </div>
-          </div>
-          <hr className="border-b-1 border-slate-200 my-8" />
-          <div className="grid grid-cols-2 gap-10">
-            <div className="col-span-1 pr-4">
-              <h2 className="text-xl font-base">Tags</h2>
-              <p>
-                Give your project tags that you believe reflect a future it is
-                building towards. These are one word tags like:
-              </p>
-              <br />
-              <p>AI, crypto, worldbuilding, storytelling</p>
-            </div>
-            <div className="col-span-1">
-              <FormField
-                control={control}
-                name="tags"
-                render={({ field }) => (
-                  <FormItem className="pb-4">
-                    <FormLabel>Enter your tags separated by commas</FormLabel>
-                    <FormControl>
-                      <Input {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </div>
-          </div>
-        </div>
-        <div className="border border-slate-200 rounded p-10 mb-10">
-          <h1 className="font-base">The Team</h1>
-          <p className="form-subheading">{"Who's building this project?"}</p>
-          <hr className="border-b-1 border-slate-200 my-8" />
-          <div className="grid grid-cols-2 gap-10">
-            <div className="col-span-1 pr-4">
-              <h2 className="text-xl font-base">Team</h2>
-              <p>
-                Please add your team members names and brief bio. Note that your
-                email will not be visible on the platform.
-              </p>
-            </div>
-            <div className="col-span-1">
-              <TeamFields />
-            </div>
-          </div>
-        </div>
-        <div className="border border-slate-200 rounded p-10 mb-10">
-          <h1 className="font-base">Support</h1>
-          <p className="form-subheading">What support are you looking for?</p>
-          <hr className="border-b-1 border-slate-200 my-8" />
-          <div className="grid grid-cols-2 gap-10">
-            <div className="col-span-1 pr-4">
-              <h2 className="text-xl font-base">Collaborators</h2>
-              <p>
-                Specify the type of collaborators you need, technical or
-                non-technical, advisors, audiences, or allies. Provide a project
-                description to invite people to assist you.
-              </p>
-            </div>
-            <div className="col-span-1">
-              <FormField
-                control={control}
-                name="collaborators"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>
-                      {"List the collaborators you're looking for"}
-                    </FormLabel>
-                    <FormControl>
-                      <Textarea {...field} />
-                    </FormControl>
-                    <FormDescription>
-                      {"Leave blank if you don't need any collaborators"}
-                    </FormDescription>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </div>
-          </div>
-        </div>
+        <ProjectSection />
+        <TeamSection />
+        <SupportSection />
         <div className="border border-slate-200 rounded p-10 mb-10">
           <h1 className="font-base">Milestones</h1>
           <p className="form-subheading">{"What's your roadmap?"}</p>
